+++ UnityProject/Saneject/Assets/Plugins/Saneject/Runtime/Bindings/ComponentFilterBuilder.cs
@@ -218,7 +218,19 @@ namespace Plugins.Saneject.Runtime.Bindings
         /// </summary>
         public ComponentFilterBuilder<TComponent> WhereTargetIs<TTarget>()
         {
-            binding.AddTargetFilter(obj => obj is TTarget, typeof(TTarget));
+            binding.AddInjectionTargetFilter(obj => obj is TTarget, typeof(TTarget));
+            return this;
+        }
+
+        /// <summary>
+        /// Filter to only resolve dependencies using this binding when the injection target
+        /// member (field or property) has one of the specified names.
+        /// </summary>
+        public ComponentFilterBuilder<TComponent> WhereMemberNameIs(params string[] memberNames)
+        {
+            foreach (string memberName in memberNames)
+                binding.AddInjectionTargetMemberNameFilter(name => name == memberName, memberName);
+
             return this;
         }